refactor(examples): 迁移 dnd 至 plugins 目录
