+++ examples/src/pages/plugins/dnd/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Graph, Node, Dnd } from '@antv/x6'
-import '../index.less'
+import '../../index.less'
 
 export class DndExample extends React.Component {
   private graph!: Graph
